requirements.txt
Pillow==9.4.0
inflect==6.0.2
requests==2.28.2